+++ src/main/java/net/inmisaddon/InmisAddonClient.java
@@ -9,6 +9,8 @@ import net.inmisaddon.util.BackpackTab;
 @Environment(EnvType.CLIENT)
 public class InmisAddonClient implements ClientModInitializer {
 
+    public static final boolean isFirstPersonLoaded = FabricLoader.getInstance().isModLoaded("firstperson");
+
     private static final boolean isLibZLoaded = FabricLoader.getInstance().isModLoaded("libz");
 
     @Override